change warning level to 'extra' and resolve warnings
Counters have been changed from `int` to `size_t`. Unused parameters
have been removed. Note that the algorithm that decides the number of
plot units to use has been changed as well (from `n / 2 - 1` to `n / 2`).

+++ src/TaskScheduler.h
@@ -75,11 +75,11 @@ namespace Luculentus
 
       /// The number of TraceUnits to use
       /// (not all of them have to be active simultaneously).
-      int numberOfTraceUnits;
+      size_t numberOfTraceUnits;
 
       /// The number of PlotUnits to use
       /// (not all of them have to be active simultaneously).
-      int numberOfPlotUnits;
+      size_t numberOfPlotUnits;
 
       /// An array of all TraceUnits in the tracer.
       TraceUnit**  traceUnits;
